+++ CHANGES.rst
@@ -8,6 +8,8 @@ Here you can see the full list of changes between each SQLAlchemy-Continuum rele
 ^^^^^^^^^^^^^^^^^^
 
 - Add optional native trigger based versioning for PostgreSQL dialect
+- Fix custom transaction column name handling with models using joined table inheritance
+- Fix subquery strategy support for models using joined table inheritance
 
 
 1.0.3 (2014-07-16)